# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "API de encurtador de links"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "paths": {
      "/url": {
        "post": {
          "description": "Encurta um link",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "link",
              "in": "body",
              "description": "O link que será encurtado",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "link": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Link encurtado com sucesso",
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "/url/{id}": {
        "get": {
          "description": "Redireciona para o link original",
          "produces": [],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "O id do link encurtado",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "302": {
              "description": "Redirecionamento para o link original"
            },
            "404": {
              "description": "Link não encontrado",
              "schema": {
                "type": "object",
                "properties": {
                  "error": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  }